tweaking-tests-for-ci
